move file to services and make copy or array
